backend/requirements.txt
Django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
python-decouple==3.1
pytz==2018.5
